fix(iostream): the logic of line extraction incorrect

+++ include/LIBFT/iostream.h
@@ -6,7 +6,7 @@
 /*   By: htsang <htsang@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/16 20:56:31 by htsang            #+#    #+#             */
-/*   Updated: 2023/06/26 02:02:04 by htsang           ###   ########.fr       */
+/*   Updated: 2023/06/26 23:11:54 by htsang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ int					ft_iostream_init(struct s_ft_iostream *iostream);
 
 void				ft_iostream_reset(struct s_ft_iostream *iostream);
 
-int					ft_iostream_read(struct s_ft_iostream *iostream, int fd);
+ssize_t				ft_iostream_read(struct s_ft_iostream *iostream, int fd);
 
 int					ft_iostream_read_until(struct s_ft_iostream *iostream, \
 int fd, t_ft_string_slice match);
